requirements.txt
streamlit==1.38.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
together==1.1.0
aiohttp>=3.9.3,<4.0.0
feedparser==6.0.11
huggingface_hub==1.28.0
transformers>=4.34.0,<5.0.0
numpy==1.26.0
